stripe customers are gross
it turns out that when a stripe customer has been deleted, the customer
object does not even define regular customer object methods like
the #subscriptions method, and calling any of those methods causes a
NoMethodError.

sadly and conversely, customers that have _not_ been deleted do not
define the #deleted method, which means that calling the #deleted method
on a customer that has not been deleted _also_ causes a NoMethodError.

:cry: